Made the new tracking component output particles so that they can be written to a text file.  ComponentOutputFile just has comments added.
Previous tracking components only assign tracks to a particle each
frame.  ComponentMinCostFlowTracking only outputs tracks every N
frames, for a potentially large value of N.(But small enough to fit in
memory.) It now outputs particles at the same time as it outputs
tracks.  The particles that it outputs are for many previous frames.
ComponentOutputFile doesn't need to be changed as long as the
particles for each track come in chronological order, which
ComponentMinCostTracking takes care to do.




git-svn-id: svn+ssh://svn.code.sf.net/p/swistrack/code@796 18c3c8b4-d610-0410-8eed-94a590f67c64